levels/03-smart-quality-detection/requirements.txt
pymupdf
openai
python-dotenv
